--- src/main/java/racingcar/domain/DefaultNumberGenerator.java
@@ -1,14 +1,0 @@
-package racingcar.domain;
-
-public class DefaultNumberGenerator implements NumberGenerator {
-    private final int num;
-
-    public DefaultNumberGenerator(int num) {
-        this.num = num;
-    }
-
-    @Override
-    public int getNumber() {
-        return num;
-    }
-}
