+++ src/main/java/racingcar/domain/NumberGenerator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.NumberGenerator;
 
 public interface NumberGenerator {
     int getNumber();
