+++ src/main/java/racingcar/domain/NumberGenerator/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.NumberGenerator;
 
 public class RandomNumberGenerator implements NumberGenerator {
     @Override
